requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
curl_cffi==0.15.0
distro==1.9.0
extra-streamlit-components==0.1.81
Flask==3.1.3
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.2
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-secret-manager==2.27.0
google-genai==1.71.0
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-oauth==0.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multitasking==0.0.12
narwhals==2.19.0
numpy==2.4.4
numpy-financial==1.0.0
oauthlib==3.3.1
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.6
plotly==6.6.0
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pyotp==2.9.0
pyparsing==3.3.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.24
pytz==2026.1.post1
qrcode[pil]==8.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
streamlit==1.56.0
streamlit-oauth==0.1.14
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.8
wheel==0.46.3
yfinance==1.2.1
